完善

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/spells/ClericSpell.java
@@ -18,7 +18,7 @@ public abstract class ClericSpell {
 
 	public abstract void onCast(神圣法典 tome, Hero hero);
 
-	public float chargeUse( Hero hero ){
+	public int chargeUse( Hero hero ){
 		return 1;
 	}
 
@@ -31,11 +31,11 @@ public abstract class ClericSpell {
 	}
 
 	public String shortDesc(){
-		return Messages.get(this, "short_desc") + " " + Messages.get(this, "charge_cost", (int)chargeUse(Dungeon.hero));
+		return Messages.get(this, "short_desc") + " " + Messages.get(this, "charge_cost", chargeUse(Dungeon.hero));
 	}
 
 	public String desc(){
-		return Messages.get(this, "desc") + "\n\n" + Messages.get(this, "charge_cost", (int)chargeUse(Dungeon.hero));
+		return Messages.get(this, "desc") + "\n\n" + Messages.get(this, "charge_cost", chargeUse(Dungeon.hero));
 	}
 
 	public boolean usesTargeting(){
