+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/spells/Smite.java
@@ -10,7 +10,6 @@ import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Buff;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.FlavourBuff;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Invisibility;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Hero;
-import com.shatteredpixel.shatteredpixeldungeon.actors.hero.HeroSubClass;
 import com.shatteredpixel.shatteredpixeldungeon.items.artifacts.神圣法典;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.Weapon;
 import com.shatteredpixel.shatteredpixeldungeon.mechanics.Ballistica;
@@ -44,8 +43,8 @@ public class Smite extends TargetedClericSpell {
 	}
 
 	@Override
-	public float chargeUse(Hero hero) {
-		return 2f;
+	public int chargeUse(Hero hero) {
+		return 2;
 	}
 
 	@Override
